namespace gis metadata client and configuration within gis
